Refactoring => discovery, cloudstore, database, gazette packages.
These live under "api-server", which will eventually be "backend"
or something similar.

"discovery" is for anything related to discovering service
configuration. Etcd support, service-agnostic endpoint management, etc.

"cloudstore" is for persistence via GCS, S3, or a similar service. We
should add an in-memory implementation for richer testing.

"database" is for aerospike and any other database adapters we may use.