Basic integration test
Define a basic integration test:

- Trigger a sync on a test pixel which should result in the linkage
  between two unique IDs (UUIDs) which are linked via a device cookie.

- The linkage emits a graph event which new queryd-testing looks at, and
  emits a row into a test stream.

- Assert that the lines emitted into the test stream are for the UUIDs
  being linked together via the device cookie.

The test lives in a test harness which should be extensible with other
suites, but does not run with 'go test' as do the current tests. The
idea is to build a standalone test runner that we can periodically run
in production and then output the passed/failed tests to a dashboard.

Issue #381