Add topic.Topic to message.Message
This requires moving around various bits to remove a package cycle. Motivation
is that this simplifies the Gazette consumer API and simplifies plumbing of
messages being concurrently consumed from multiple topics.

As topic.Sink is now an interface type with few dependencies, we're able
to remove the TopicSink interface plumbing which has popped up in a
number of packages.

Issue #1197