Wait for queue to empty out before exiting.
Define a shutdown behavior for the StartPersisting() loop.

Normally, StartPersisting() will do a wait <-> converge loop.

Define a Stop() method that activates a shutdown behavior where we
stop respecting the interval, begin warning on new Persist() requests,
and we continuously converge until the queue empties out (or we get
SIGKILLed by Kubernetes after 6 hours, which hopefully we never get to.)

Once the queue is observed to be empty whilst in shutdown mode, signal
to Stop() that we are done, allowing the process to exit.

Issue #1295