Create visualization of Gazette flow metrics
Made with d3.js and its Sankey plugin, shows the flows between readers
and writers of journals. The size of each flow represents the amount of
data flowing through. Readers with lag are highlighted in various shades
of red.

The data the visualization uses is a compilation of debug/vars output
from all hosts in the infrastructure. The service takes care of keeping
running totals of heads and read/write rates for every (journal, actor)
pair.

This functionality is baked into a new API server, varzd. The static
assets required to view the output in a browser are actually embedded
into the Go binary, and served using a http.FileServer within. So, the
idea is that we deploy this to production, port-forward to it, then just
navigate to that forwarded port in a web browser.

This is the temporary solution until we have many more debug services
and have a good idea how to gate them behind proper API authorization.

Issue #1246