varz count and average tagging for datadog
Rather than put all our metadata in the name, allow specially formatted
varz metric names to reduce into a shorter name with a number of
key-value tag pairs to be submitted with our statsd integration.

For example, sync.pixelIn.1234 becomes sync.pixelIn.#pid.1234,
which processes to a statsd gauge of varz.sync.pixelIn with 'pid=1234'
as a tag (same kind of tag as hostname and pod.)

Any arbitrary number of KV pairs is accepted. The formatter is
forgiving, so if it cannot transform a KV pair into a tag for any reason
(minimum length of metric name, key missing value, etc.) it just leaves
it in verbatim. The whole shebang remains compatible with traditional
varz console logging, though in some cases I added informational
parameters for group-bys that weren't there before.

The idea is in DataDog you can then log everything around the
sync.pixelIn metric and do breakdowns by pid if you so desire. Further
down the line, it's plausible that we can do breakdowns by seller AND
pid based on pixel model metadata, but this would make console logging
unwieldy.

Issue #1265 (part of it)