Remaining `reporting` components
Consumer performs the heavy-lifting of updating a relation from input
topics, as a Gazette Consumer.

DBScanner and GatherScanner are implementations of the Scanner
interface:
 - DBScanner directly executes a Query against a local RocksDB
   which represents a Shard of the relation.
 - GatherScanner executes a Query by identifying the current set of
   Relation masters, and then pushing down (scattering) the Query to each
   Shard for execution. It presents a Scan interface which incrementally
   streams and combines the results.

ScanAPI is the server-side counterpart to GatherScanner, which provides
an HTTP API for executing a Query against a mastered Shard.

`reportingd` is the principal production binary for building and serving
relations. It executes a configured model Relation as a Gazette
Consumer, and wires a ScanAPI for serving results from
currently-mastered Shards.

`query-relation` is a command-line tool for executing a Query against a
live relation, and viewing results as a TSV.

Also add ReportingSuite, which implements end-to-end tests over
reporting components in the context of a word-counting relation.

Bonus: fixup EventProcessorSuite from prior topics pkg merge/rebase.

Issue #1354, #1355