Factor topic consumption into groups.
Until now, the v2 consumer assumed that all topics you were consuming as
part of a consumer had to be consumed in lockstep. This disregarded the
reality that we often like to consume from multiple topics, but don't
care about scanning across topics (aligned partition assignments, etc.)

To allow this, Topics() returning []*topic.Description is now Groups()
returning a TopicGroups structure where each entry represents a group of
topics that are to be considered as a unit. To consume two topics
independently from one another, just create two topic groups.

Each group must be named and the consumer must return groups whose names
are in lexical order. This prevents fallout from occurring if we choose
to delete a group from the listing.

Issue #1402