Switch `gazette` service over to consensus.Allocator
Context => Runner, an implementor of `consensus.Allocator`

Runner receives and forwards route change signals on to Router, which
manages the Journal routing table and local replica state. Journal names
and route endpoints use URL query encoding, and Journals must have a
successful transaction under the current route topology before replicas
may be promoted.

Add CreateAPI, which handles POST of a Journal and creates the
corresponding Etcd item. Races between Journal creation, and attempted
use of the journal is a bit of a annoyance: clients often immediately try
to use a created journal, which won't be ready until the distributed
allocation of replicas has completed. Try to account for this by
watching the item in etcd and waiting for "ready" values from all
replicas. It helps a lot, but is not perfect.

Update ReadAPI and WriteAPI with handling for ErrNotExist, to send
the RouteToken as a header, and to use `journal`'s protocol error
=> HTTP code mapping.

Use Explicit journal package import in tests (otherwise, the new
journal.Client conflicts with gazette.Client).

Issue #1544