better `recoverylog` Playback and Recorder offset handling
Recorder now issues an initial write barrier to ascertain the current
WriteHead. This decouples the previous extant dependency on Playback
providing a correct initial offset. It also retains a recent pending
write, and will update the FSM Offset from its WriteHead at completion.
Together, these give Recorder a self-contained capability to correctly
increment the hinted FSM Offset.

Playback now issues a write barrier to determine a minimum WriteHead on
Play start, and again on MakeLive. This provides the strongest guarantee
that all prior recovery log content has been consumed, as write barriers
are transactional (as compared to Reads, which are not and could return
a stale WriteHead). It also implicitly creates the log if needed.

Issue #1544