gaztee and gaztail utilities
These utilities provide a bridge to use gazette as a bytestream from the
command line in ways familiar to shell users.

- gaztail <journal> [-n bytes] [-f]: tail the last <bytes> of a journal,
  by default 1 megabyte. -f to follow. Uses environment variables or
  -gazetteEndpoint for configuration.

- gaztee <journal>: reads stdin and writes it to the journal of your
  choice. Configured in same fashion. Arguably this could be a curl
  wrapping script but I'm really happy to have a binary that I know
  behaves the same way as the real client and absorbs all the same
  configuration variables.

This required a tweak to the Gazette client for Write requests. The
server doesn't actually care about clients passing in Content-Length on
payloads, so do not fail if the Content is not seekable.