Switch to Segment-based FSM hinting
Formerly, FSMHints indicated a starting recoverylog offset & checksum, a
sequence of expected authors, and hints as to which Fnodes may be
skipped. The downside of this approach is that the oldest, still-alive file
in the log dictates the initial read offset, and can force substantial
amounts of irrelevant log to be read.

Switch instead to including a full manifest of live Fnodes within FSMHints.
Each HintedFnode captures its covering log Segments, and FSM & Player now use
those hints to identify segments of log to process (and segments to skip!).

Tested:
  * Pre-existing `recoverylog` & `consumer2` integration tests.
  * Manual testing on local stack, with reportingd restarts & recoveries.

Release impact:
  * FSMHints are incompatible with former FSMHints.
  * Former extant FSMHints will be treated as zeroed hints.

Issue #1698