Use cgzip.Reader instead of gzip.Reader for fragment decompression.
When fetching cloud storage fragments (only; direct Gazette
broker fetches are never compressed) we need to eschew use of the
built-in, pure-Go gzip library and use cgzip as it is becoming a
throughput bottleneck.

To accomplish this, specify Accept-Encoding: gzip manually; this
disables the transparent net/http decompression on a per-request basis.
Then, on the returned response.Body, if the received encoding is
actually gzipped, wrap the Body in a cgzipReader.

Issue #1683