consumer2: Do not assume earlier ticks will be duplicated (#1841)
This is fixing a stall observed in production, where consumer transactions
appear hung waiting for minQuantumElapsed ticks which never arrive.

A suspected reason is for NTP clock adjustments, which could cause a
tick Reset by a transaction to technically arrive *before* that
transaction began. Our current handling would be to ignore such a tick,
in which case we'll never see another.

Rework the loop such that every tick which occurs within a transaction
results in a Reset, which guarantees that another tick will eventually
arrive (and we won't block). Do this by fixing delayed ticks to |txBegin|.

Tested: integration tests & local shadowrun

Issue #1838