batch_download.py: don't decompress and recompress when fetching full fragments
We must decompress and recompress when downloading a fragment starting
at a given offset (e.g. skip_delta>0). Otherwise, we can download the
full fragment.

To enable this, the outer fetch_some() loop now uses a plain file
object, and in certain cases, this plain file object will be wrapped
with a gzip.GzipFile.

Report the number of uncompressed bytes read when reading gzipped data
directly, rather than the raw bytes read off the socket. We need to
trust the fragment name for this purpose, as it is not necessarily clear
how many bytes are contained within a compressed payload.

Testing done:

- Take a reference METADATA from a random partner
- Run the old version and new version in separate directories
- zcat all the .gz files together in each dir, minus the very last file
  (which will be the broker-direct read), and compare the two
  concatenated files.
- The contents were the same, which validates this change.

Resolves #1849