URL generator for export backfill jobs.
Takes as input an Etcd prefix containing markers from each consuming
shard about the journal and offset the shard was processing at the time
it first processed this query.

Generates signed cloud storage URLs, for use with `exportd -backfillFrom
URL -query /some/specific/query -exportType some_query_type` (e.g. to be
dispatched with Cloud Dataproc or some other job-mechanism.)

As the marker specifies a maximum offset to consume up to, the operator
can also specify a minimum (Unix) timestamp of fragments to consider.
The signer will skip any fragments which have a Last-Modified header
prior to this timestamp. This is especially useful if the query in
question already has a MinimumTimestamp filter.

Next up, actually dispatching and moderating the success of this job now
that all the data is available for generating the commands we need to
perform any backfill.

Bonus: journal.Fragment.AsDirectURL refactored so that explicit duration
may be passed in, rather than having it hardcoded as time.Minute.

Issue #648