Record distribution events directly to Gazette on dataproc jobs.
It's not ideal, but allows us to process backfills requiring
distribution reporting without too much work.

The ideal solution would be to emit journal writes as Hadoop outputs,
then use a combiner and reducer to emit files that can be directly
inserted into Gazette as fully complete fragments or distributed to
a partner.

Also, allow reading query from direct JSON file. Example:

./export-backfill-launcher -gazetteEndpoint 10.60.0.4:8081 \
        -query radiumone.json -queryType match -cloudFS 'gs://' \
        -endOffsets offsets.json -minimumTimestamp 1470700800

Bonus: Fix behavior of FragmentsInRange when hitting an offset range
that does not have a persisted Fragment. A Fragment object is still
returned as part of the HEAD, but it will not have an attached URL
pointing to the persisted fragment. FragmentsInRange should break the
fragment-gathering loop on this condition.

Bonus2: PR fixes for #2302

Issue #2073