Addres Gazette broker contention contention on routesMu.
|gazette.Router.routesMu| protects not only the |routes|
map[journal.Name]*journalRoute, but its underlying values as well. This
causes lots of contention on routesMu, when it isn't strictly required
for correctness.

Most of the access patterns of |routes| are to retrieve a route, then
perform read operations on its fields or call methods on interface
pointers. For this purpose, we can use a copy of the route that's
obtained while the lock is held, and release the lock right away after
obtaining that copy.

There is one use case (Append) where a field in the journalRoute object
needs to be updated. In this case, when we need to update it, we
reacquire the lock, redo the lookup in the map, and write directly to
the underlying object. This relies on the property that routes are never
deleted out of |r.routes|, but could later be refactored.

Issue #2604