vaulted IDs, published message validation, and pruning
Add `Validate` method to PixelEvent which is invoked by SimplePublisher.
Validate changes that the event is well formed, and (in the future) that
it doesn't include sensitive identifiers we don't wish to log.

Add "pointer" IDs which stand in for sensitive IDs in the graph, and
extend GraphNode with an optional dereferenced `RawId`.

Add pruning implementions for Links and Segments, which clear out
tombstone records.

Fix Id.SimpleString, which had incorrectly included a namespace in all
cases.

Issue #2163