[PUB-2947] Add Etcd3 store for Consumer2 offsets and hints (#60)
Offsets and Hints are now written to both the v2 and v3 etcd stores
simultaneously. They are loaded from v3 first, and then from v2 if none
are found. Consumer2.Runner now has an Etcd3 property which holds a
pointer to an etcd3.Client. The various users of Runner have been
updated to set this properly. Tests have been updated to include checks
for V3 as well as V2.