Jazette (#117)
Jazette is a hackweek project with the goal of being able to easily implement a gazette consumer using Java application code.  The idea is that we should have a framework set up to allow us to easily pass messages across applications written in Go and Java.

Jazette has two main components: the Go consumer (jazetted) and the Java library.  The Go consumer is a standard, stateless Gazette consumer implementation that receives a list of topics to consume via flag.  The Java library is an interface with a single `consume()` method responsible for interpreting messages sent to it, and plumbing errors upstream.

The two components live side-by-side in a kubernetes pod and communicate via IPC using two named pipes, one in each direction.  Messages are sent as byte arrays with a length header, so it's the responsibility of the implementing Java application to interpret these messages, though the library will be responsible for splitting up the stream.

From an organizational perspective, the Jazette java code lives in the `MasterRepos` org as is typical to LiveRamp style.  The library is imported into `workspace` via submodule, and built / deployed using `workspace`'s kubernetes tooling.  The java build is also wrapped, but requires us to enable networking on our build step, since the jazette build utilizes maven, which pulls from remote sources.  TBD whether or not we keep this here since for the time being, this is just an example and not actually used in production.

Note that due to the nature of hackweek projects, this code was never deployed to production, so any future users should keep that in mind when using this.  There's a high likelihood of undiscovered bugs and performance issues that will come up when trying to scale this to a production workload.