PUB-2908 Endpoints in SQL (#142)
Deprecate the old style cloudstore.Endpoint in favor of an interface, and SFTP/S3 implementations that have a much easier to use API for connecting and testing.  Create postgres tables for these, so platform folks can easily define the endpoints, create Go and model builder plumbing to get the models into etcd.

In this CL, postman and cfs-sync are still using old-style endpoints.  A TODO is to come up with some sort of bridge compatibility while we migrate all existing endpoints to this new style, most of which will happen in [this ticket](https://jira.liveramp.com/browse/PUB-3638).

Add all the necessary model-builder plumbing and various validations, with testing on a local stack.