PUB-3705: cloudstore SFTP write buffer (#165)
Use a buffer under the sftp file Write method. The underlying
implementation of an sftp file io.Writer sends an SFTP packet and waits
for an ack every time the Write method is called, which is called
liberally by io.Copy, which in turn is used by cfs-sync and postman.
Putting a buffer on top of this greatly reduces the number of network
round trips we have to make to write a file.

Tested by adding some varz debugging to the sftp package (not
committed) and measuring average bytes and network calls per minute:

Before: 978662 bytes, 7843 writes
After: 18456022 bytes, 564 writes

... a per-network call throughput improvement of 26389%

Bonus: Don't close write closers twice in postman in the case of no
compression.