PUB-3433: Automatic endpoint validator (#216)
Scans /models/endpoints and tests new entries by attempting
to write (and then remove) a file called to the remote directory. The
name of the file is configurable on the endpoint; if no filename is 
specified, the test will only attempt to connect and read from the
directory. Will alert Slack if any failures are encountered.