Endpoints.validator fixes (#243)
* Bump aws-sdk-go submodule

Need to get the following commit in so that the endpoint validator can
successfully send an empty body when uploading to the test file:

https://github.com/aws/aws-sdk-go/commit/db7b69949079da41d74f9c6bfb150becb48db1e1

* Change file flags on enpoint-validator testfiles

RDRW is not supported in our cloudstore S3 implementation, so we have to
open them as write only. Using the same flags as we do in postman /
cfs-sync to create files for consistency, and making the change across
to the sftp implementation as well.

Bonus: change error message on close to "could not flush file" since the
upload step on multipart uploads for s3 actually happens when Close() is
invoked.

* endpoint-validator updates slack on changes

When we validateOne, it's because someone's updated a model. Always add
a positive confirmation if the validation succeeds.

* endpoint-validator default runs every 24h