consumer: Add `Client` and "primary" state.
`consumer.Client` is a building-block for interacting with consumers
which present GRPC APIs. It monitors the topology of a consumer,
maintains warm connections to each endpoint, and provides facilities for
routing to the responsible primary endpoint of a given partition.

Also add a "primary" state in Etcd, which is set by a Shard replica which
has fully recovered the Shard DB and transitioned to actively serving it.

PUB-3888