PUB-3913: SFTP File Exists Error Mapping (#330)
It looks like the SSH lib for Go is throwing a weird error mapping for
SSH file exists errors, but we do need to be able to catch these. Using
the generic SSH failure error as a file not found may possibly mask
some actual generic failures in the future, but for now this is the
quickest solution.