gazette client: likely fix for failure to invalidate cache
In production, we're seeing cases where journal routes are not invalidated
promptly in response to network errors. I _suspect_ this is because
invalidation was using the Path provided by the Location redirect, and
escaping introduced by the HTTP stack _could_ make these vary slightly
from the key that was actually used.

Update, and add a test.