recoverylog: make playback hand-off fast & deterministic
Formerly, `recoverylog` playback used behavior of waiting for a timeout
during which no new content became available from the broker. On
reaching that timeout, playback was assumed to be complete and the
consumer would switch from replicating to recording new state changes.

This invited race conditions, where a prior recorder which was delayed
in flushing final writes to the log would race with the new recorder.
Ordinarily FSMHints are sufficient to resolve the forked log history,
but this is a problem for other warm-standby replicas which are
live-tailing the log: depending on which raced write wins, they may
follow a fork of log history which is shutting down, and ignored the new
recorder entirely.

Fix this by changing player behavior to allow for injecting "hand-off"
noop RecordedOps, which serve as write-barriers that effectively give
control of the log to a new recorder in the eyes of any other tailers of
the log. Timeout-based polling is removed entirely.

This has a number of spill-over implications. Playback must now use a
mix of blocking and non-blocking reads, and `RetryReader` is refactored
to support this behavior. Use of Context is introduced to signal
deadlines and cancellations, and the server-side blocking duration
mechanic is deprecated (in favor of client-side cancellation). The
Author ID must be shared between the `Player` and `Recorder`, motivating
a refactor of `consumer/master.go`.

Also, switch ShardPreInitHook => ShardPostInitHook. In usages such as
ShardIndex, the expectation is the Shard is fully initialized, making a
Post-hook appropriate.