Revert "consensus: ignore routes which expire in the near future"
This reverts commit b2c22c69e59650864d18e39999908fb0af42c0c5.

This change was found to cause seemingly harmless races in the
consensus.Allocator, where a restarted process would race it's
own prior locks (which it was ignoring).

It will also cause client applications to route traffic to replicas
which don't yet believe themselves master, which is arguable behavior.

Given the subtle data races this introduces, it just seems a bad idea.
Better to let Etcd be the sole arbiter of when data has expired.