Initial, basic Helm charts for standing up Gazette
Very much a work in progress. This includes a Helm chart for Gazette
itself, with dependencies on Minio & Etcd, and a Helm chart for the
`word-count` example.

Minor updates to `cloudstore` and `index_watcher.go` are included to
support Minio work around a difference in directory S3 semantics (vs
S3).

test/NOTES.rst includes more detail on how to use these in standing up a
functional Gazette cluster & application on a minikube stack.