run-consumer: use Viper for configuration
As opposed to supplying all arguments by flags. The intended model is
that applications will define K8s ConfigMaps which package their
configuration, and further override elements of that configuration
through environment variables.

The `word-count` example is refactored with this model, where a YAML
ConfigMap is specialized by environment overrides.