examples: add `stream-sum` example, build, and chart
This example consists of a "chunker", which emmits sequenced chunks of
concurrent data streams, and a "summer", which observes chunks and tracks
ongoing CRC64 sums of their contents, eventually writing the final sum
into a "sums" Journal at stream completion. The chunker, meanwhile,
observes the final sum and compares with its expectation (based on chunks
it emitted).

stream-sum is a stand-alone example, requiring no additional user input
to function, and also an integration or "smoke" test, as it explicitly
checks for and fails if at-least-once semantics are violated in either
the broker or consumer, or if data corruption occurs.

This example would be a good candidate for a "chaos monkey" type
apparatus, as it should not fail & make progress in spite of soft
or hard failures in brokers or consumers. It could also become a
baseline for performance benchmarking.