recoverylog: refactor & remove FSM.LogMark
FSM had no business representing a notion of the current "offset",
which was either a real offset (in the case of playback), or a somewhat
recent write head (in the case of recording). The original need was so
that FSM could track metadata for building hints that allow a player to
get fairly close to the actual log offset they need to begin reading
from.

With RecordedOp.First/LastOffset, the FSM LogMark is now totally
duplicative. Remove it. This has some cascading implications:
  * Recorder needs to explicitly track the writeHead it's using as an
    approximation for operation FirstOffsets.
  * Playback similarly needs to track it's own mark.

During playback, we also want to prove that computed LastOffsets actually
match what we're consuming, which required surfacing the RecordedOp
itself from playOperation. Also take a stab at further breaking up
playOperation (into applyOperation & enactOperation) in the name of
readability and explicit-ness. Lots of other comment updates.

Issue #26