consumer: maintain GRPC connections to all PRIMARY endpoints
Even if they're not explicitly listed in ConsumerState.Endpoints. This
allows clients to talk to PRIMARY shards hosted by members in the
process of shutting down. That shut-down could take a while, as shards
are recovered by other processes, and clients should be able to access
shards in the meantime.

Issue #50