protocol: add dispatcher balancer and LabelSet refactor
dispatcher is a GRPC load balancer which is aware of protocol.Route
semantics. It allows a tight integration between grpc.SubConn connection
management and the Routes which may be known to the application (either
through caching, or because the application has a view on etcd state
which implies routing). Routes are attached to the Contexts of GRPC
client requests; the dispatch balancer extracts the Route and uses
current connectivity state to select an appropriate SubConn to which the
request is dispatched. It is zone aware, and will prefer intra-zone
endpoints.

Refactor LabelSet to be a multi-map. Set operations are still defined
over label keys (eg, when intersecting, all values of a label must match).
Also add routines for parsing and emitting selector strings (see
ParseLabelSelector).

Rename Broker service => Journal service, and Update* => Apply*.

Include other small tweaks from fixing -race test failures.

Issue #46