client: protocol dispatcher, AsyncAppend dependencies, & PolledList
Remove Dialer and RoutingClient. Add RouteCache, which deals only
with concerns of Route caching and is a protocol.DispatchRouter. Update
and clarify interfaces for RoutedJournalClient, AsyncJournalClient.

Update AppendService to track priors of an Append being started.  These
are AsyncAppend operations which must complete before the current append
operation can be started. This mechanism can be used to build comprehensive
cross-journal write barriers.

Add ListAll and PolledList for periodic enumeration of JournalSpecs
matching a selector.