recoverylog: copy and modify for v2
Much of the playback player machinery is simplified with the v2
RetryReader API. Playback includes some additional proactive error
checks around hints, and the Player interface has been altered to
better fit V2 usages.

Recorder is adapted to `client.AsyncJournalClient`, and is also
decoupled from being specific to RocksDB. A recorder of an afero.FS
interface is included; any DB which implements to this interface is
automatically recordable (future work: this could be used with
https://github.com/dgraph-io/badger/).

Recorder distinguishes between "weak" and "strong" write barriers.
The former implements v1 behavior of waiting until the recoverylog
has committed. The latter is new behavior which uses
client.AsyncAppendService's dependecy management to require that
_all_ pending journal writes complete before the barrier returns.