broker: re-work health checks and offset-jump behavior
Introduce an AppendRequest.Offset field, and if present, require that it
exactly match the furthest known journal offset. Change broker behavior
to now refuse Append RPCs where the pipeline's offset is less than the
fragment index's maximum offset, _unless_ the AppendRequest.Offset
matches the index offset. Only in this case is the pipeline offset
rolled forward to the index offset.

This behavior precludes a class of failures we've seen with Gazette V1,
whereby replication consistency is lost (somehow) and a new topology of
brokers start up. They take some recent fragment offset and being
commits from that point; meanwhile previous brokers upload fragments and
the net effective is that a range of offsets are written twice.

With V2, an operator must get involved and manually "jump" the pipeline
offset forward if consistency is lost, after verifying that all
prior brokers have exited & fragments are uploaded, and that no further
data can possibly be forthcoming. The plan is to add a `gazctl` command
to simplify this task.

A secondary benefit of AppendRequest.Offset is it can be used to
construct idempotent, retry-able append operations in contexts where one
knows they are the exclusive writer to a journal. It also provides a
basic transactional primitive.

Finally, refactor replica.go and extend with better tests.

Issue #68