recoverylog: fix Peek(1) flake
playOperation expects that, upon Peek() returning, an entire frame can
be read without blocking. Usually that's true, but a test flake
demonstrated that under certain conditions of garbage data, Peek(1)
could return when there was not yet an complete frame, causing playback
to fail. Update to Peek(message.FixedFrameHeaderLength), which
guarantees that an entire frame can be read (even if it's garbage).

Also decorate error handling with better breadcrumbs for tracking down
error contexts, and switch to using client.Append when determining the
log head.