Leverage cached image layers in CircleCI
The image liveramp/gazette-base:2.1.0 is an up-to-date build of the
"base" stage. Most importantly, it allows subsequent builds to skip
building RocksDB.

For unknown reasons, to get builds of later stages to use cached layers
on CircleCI, a "--cache-from" option for each referenced earlier stage
is required. This is not required on local runs, but still works.