brokertest/consumertest: Apply in a goroutine, and instrument idleCh sends
There's a rare deadlock where the server's WaitForRevision in the Apply
handler blocks on KeySpace.RLock(), which blocks on the Etcd Watch being
applied, which blocks on TestHook's send to |idleCh|, which is blocked
because brokertest/consumertest are not yet reading from |idleCh|. So,
perform the Apply in a goroutine.

Also instrument TestHook to panic after being stalled on send for a few
seconds. That should make this class of deadlock very obvious to identify.