Use more specific tag for base image
The tag "1.11" is expected to change regularly. When it does,
liveramp/gazette-vendor:latest is no longer valid as a cache source for
CI builds and results in much longer build times. Including the patch
version makes the tag more stable while adding the Debian release
differentiates it from Alpine-based images.