broker: fix subtle test context cancellation race
Both teststub.NewLoopbackServer() and stream opened by the test were
monitoring `tf.ctx` for cancellation. Usually the stream "read" the
cancellation first, but it's also possible that the connection is closed
prior to the stream selecting `tf.ctx.Done`, in which case a different
error is returned ("transport is closing"). Fix by wrapping with another
context & cancelFn for the stream which is explicitly cancelled, rather
than using cleanup().