Support fragment store URL rewriting
When migrating from multiple Gazette clusters to a single one, it may be
desirable to not move existing fragments. However, this both forces the
name of the journal in the target cluster and may result in a name
conflict between two journals with the same name but different fragment
stores.

The "find"/"replace" query parameters implemented on "file", "s3", and
"gs" backends provides a workaround that avoids moving fragments.

For example, if previously there was

- `foo-topic/part-000` backed by `s3://bar-journals/`
- `foo-topic/part-000` backed by `gs://bar-journals/`

the two would collide in the target cluster despite being distinct
journals (unless one was renamed).

The example below shows how to give the journals different names in the
target cluster while still being backed by the existing fragments.

```
  name: "s3-foo-topic/"
  stores:
    - "s3://bar-journals/?find=s3-foo-topic&replace=foo-topic"
  children:
    - "s3-foo-topic/part-000"

  # ...

  name: "gs-foo-topic/"
  stores:
    - "gs://bar-journals/?find=gs-foo-topic&replace=foo-topic"
  children:
    - "gs-foo-topic/part-000"
```