consumer: two breaking ShardSpec changes & one API change
 1) ShardSpec.HintKeys => HintPrefix & HintBackups
 2) ShardSpec.RecoveryLog => RecoveryLogPrefix
 3) BeginFinisher.FinishTxn adds an error argument.

1) Formalize the exact specification of HintKeys in Etcd, as a composite
between a HintPrefix, the Shard ID, and primary or backup versions of
the hints. This significantly narrows the space of allowable hint keys,
but doesn't hinder any use case we're aware of and is likely to be a lot
less confusing to operators.

This allows HintPrefix & HintBackups to be held common across a collection
of ShardSpecs, which makes it possible to write tooling that avoids DRY.

2) Similarly, formalize that recovery logs must be a composite of the
RecoveryLogPrefix and Shard ID. Again, this narrows the space of allowable
recovery logs, but only along the lines of reasonable best practice. Like
HintPrefix, it allows a collection of ShardSpecs to reasonably share a
single RecoveryLogPrefix, which can greatly reduce stutter in tooling.

3) Surface a consumer transaction error to FinishTxn. Implementations
of 2PC require this, because a FinishTxn as part of an errored transaction
must know not to, for instance, write commit acknowledgements of a commit
that didn't actually happen.

This is a breaking change, and required updates to a number of tests
and the examples.