charts: make extraArgs a map with value templates
Instead of a big template string. This allows granular extra arguments to be
specified at any point, and Helm will do the heavy lifting of merging into a
combined representation, using a last-in-wins strategy. Treat argument values
as templates to be executed, preserving the desired dynamicity of values.

Remove the "config" ConfigMap. No clear use-case exists for this in server
deployments (it's far more useful for developer-run CLI tooling). Deployments
should prefer explicit arguments and environment variables, though they're free
to provide and mount a ConfigMap if that's truly desired.