Decentralize adding of commands
With the number of commands growing and the volume of text in the
description of each, `main.go` was becoming difficult to read.
Implementing each command in a file (pre-existing) and adding it in an
`init` function (new change) places command documentation closer to its
implementation and is considered idiomatic by
github.com/jessevdk/go-flags.