allocator: add "Overflow" node to network structure
The allocator flow-network scales member capacities via calculated zone scaling
factors, which are designed to encourage appropriate balancing of items across
members.

But, we've seen edge cases with this scaling where a larger number of items does
not fully allocate across a smaller number of members, each having large
capacities. I suspect these cases are impossible to detect ahead of time,
without actually running push/relabel.

The core problem is we're artifically limiting the capacity of members in the
flow network. This PR undoes that limitation but preserves the incentive to
allocate items evenly across members, by introducing an "Overflow" node. Each
Member node having remaining capacity beyond it's scaled capacity, expresses
that excess via an arc to the Overflow. But, the Overflow node is initialized
with a large height: in fact, the precise height which makes it the last
alternative for dealing with flow excess before that excess is instead pushed
back to the Source (terminating the algorithm).

Issue #157